Relocate components to only one components folder outside app folder
